dev/okay: organize lib, add metrics constructors, make UniqueKeys required
