+++ dev/okay/okay.go
@@ -11,76 +11,12 @@ import (
 	"io"
 	"net/http"
 	"sync"
-	"time"
 
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 	"github.com/sourcegraph/sourcegraph/lib/log"
 )
 
 var okayhqAPIEndpoint = "https://app.okayhq.com/api/events/v1"
-
-// Metric represents a particular metric attached to an event.
-type Metric struct {
-	// Type is either "count", "durationMs" or "number".
-	Type string `json:"type"`
-	// Value is the actual value reported for this metric in a given event.
-	Value float64 `json:"value"`
-}
-
-// customEvent represents a custom event sent to the OkayHQ API.
-type customEvent struct {
-	// Event is the event type, should be always set to "custom".
-	Event string `json:"event"`
-	// Name is the custom event name, used to select those events amonst others to build dashboards.
-	Name string `json:"customEventName"`
-	// Timestamp is the time at which this event occured.
-	Timestamp time.Time `json:"timestamp"`
-	// Identity ties this specific event to a particular user, enabling to filter events on various group predicates
-	// such as teams, organizations, etc ... (optional).
-	Identity *eventIdentity `json:"identity,omitempty"`
-	// UniqueKey lists the property keys that are used to uniquely identify this event (optional).
-	//
-	// Sending another event with the same UniqueKey results in overwritting the previous event,
-	// enabling to replay events with historical data or to correct incorrect events that were previously sent.
-	UniqueKey []string `json:"uniqueKey,omitempty"`
-	// Metrics are a map of okayMetric whose keys are the metric name.
-	Metrics map[string]Metric `json:"metrics"`
-	// Properties are a map of additonal metadata (optional).
-	Properties map[string]string `json:"properties,omitempty"`
-	// Labels are a list of strings used to tag the event.
-	Labels []string `json:"labels,omitempty"`
-}
-
-// eventIdentity represents the identity to attach to an event.
-type eventIdentity struct {
-	// Type represents from where this identity is registered, should always be "sourceControlLogin".
-	Type string `json:"type"`
-	// User is the unique identifier to reference this identity amongst its Type.
-	User string `json:"user"`
-}
-
-// Event represents a custom Event to be sent to OkayHQ.
-type Event struct {
-	// Name is the custom event name, used to select those events amonst others to build dashboards.
-	Name string
-	// Timestamp is the time at which this event occured.
-	Timestamp time.Time
-	// OkayURL is used to generate a clickable link in OkayHQ's UI when browsing this event.
-	OkayURL string
-	// GitHub login this event is attached to (optional).
-	GitHubLogin string
-	// UniqueKey lists the property keys that are used to uniquely identify this event (optional).
-	//
-	// Sending another event with the same UniqueKey results in overwritting the previous event,
-	// enabling to replay events with historical data or to correct incorrect events that were previously sent.
-	UniqueKey []string
-	// Properties are a map of additonal metadata (optional).
-	Properties map[string]string
-	// Metrics are a map of okayMetric whose keys are the metric name.
-	Metrics map[string]Metric
-	// Labels are a list of strings used to tag the event.
-	Labels []string
-}
 
 // Client collects and submit metrics to the OkayHQ custom events API and is safe to use
 // concurrently.
@@ -161,6 +97,9 @@ func (c *Client) Push(event *Event) error {
 	}
 	if len(event.Metrics) == 0 {
 		return errors.New("okayhq: event must have metrics")
+	}
+	if len(event.UniqueKey) == 0 {
+		return errors.New("okayhq: event must have unqiue property keys")
 	}
 	for _, k := range event.UniqueKey {
 		if _, ok := event.Properties[k]; !ok {
